# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinical Review Copilot API (P-194)",
    "description": "Backend API for Clinical Review Copilot platform — FHIR R4 & PDF/OCR Ingestion",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ingestions": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest File",
        "description": "Ingest a PDF, Image, or JSON document for a patient.\n\nIf patient_id is not provided, AI will parse the document to match an existing patient,\nor automatically create a new patient with extracted demographic data.",
        "operationId": "ingest_file_api_v1_ingestions_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_file_api_v1_ingestions_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionBatch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "List Ingestions",
        "description": "List recent ingestion batches.",
        "operationId": "list_ingestions_api_v1_ingestions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngestionBatch"
                  },
                  "title": "Response List Ingestions Api V1 Ingestions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ingestions/{batch_id}": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Get Ingestion Status",
        "operationId": "get_ingestion_status_api_v1_ingestions__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionBatch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/process": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Process Patient",
        "operationId": "process_patient_api_v1_patients__patient_id__process_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPatientRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPatientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_ingest_file_api_v1_ingestions_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "new_patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Patient Name"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingest_file_api_v1_ingestions_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestionBatch": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "source_document_id": {
            "type": "string",
            "title": "Source Document Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "source_checksum": {
            "type": "string",
            "title": "Source Checksum"
          },
          "received_at": {
            "type": "string",
            "title": "Received At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "data_watermark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Watermark"
          },
          "counts": {
            "$ref": "#/components/schemas/IngestionCounts"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/IngestionErrorDetail"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "status",
          "format",
          "source_document_id",
          "source_checksum",
          "received_at"
        ],
        "title": "IngestionBatch"
      },
      "IngestionCounts": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted",
            "default": 0
          },
          "quarantined": {
            "type": "integer",
            "title": "Quarantined",
            "default": 0
          },
          "needs_verification": {
            "type": "integer",
            "title": "Needs Verification",
            "default": 0
          }
        },
        "type": "object",
        "title": "IngestionCounts"
      },
      "IngestionErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "IngestionErrorDetail"
      },
...
      "ProcessPatientRequest": {
        "properties": {
          "profile_versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Profile Versions"
          }
        },
        "type": "object",
        "title": "ProcessPatientRequest"
      },
      "ProcessPatientResponse": {
        "properties": {
          "process_id": {
            "type": "string",
            "title": "Process Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "data_watermark": {
            "type": "string",
            "title": "Data Watermark"
          }
        },
        "type": "object",
        "required": [
          "process_id",
          "status",
          "patient_id",
          "data_watermark"
        ],
        "title": "ProcessPatientResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}